all of the tests pass, activity service tests finished

+++ activity-microservice/src/main/java/nl/tudelft/sem/template/example/controllers/ActivityController.java
@@ -32,14 +32,14 @@ public class ActivityController {
 
 
     @DeleteMapping("/deleteId/{id}")
-    public void deleteById(@PathVariable Long id) throws UnauthorizedException, ActivityNotFoundException {
+    public void deleteById(@PathVariable int id) throws UnauthorizedException, ActivityNotFoundException {
         NetId username = new NetId(authManager.getNetId());
         activityService.deleteById(username, id);
     }
 
 
     @PutMapping("/edit/{id}")
-    public void editActivity(@PathVariable Long id, @RequestBody ActivityRequestModel request) throws UnauthorizedException, ActivityNotFoundException {
+    public void editActivity(@PathVariable int id, @RequestBody ActivityRequestModel request) throws UnauthorizedException, ActivityNotFoundException {
         NetId username = new NetId(authManager.getNetId());
         activityService.editActivity(username, id, request);
     }
@@ -77,7 +77,7 @@ public class ActivityController {
     }
 
     @GetMapping("/user/{id}")
-    public NetId getOwnerById(@PathVariable("id") long id) throws ActivityNotFoundException {
+    public NetId getOwnerById(@PathVariable("id") int id) throws ActivityNotFoundException {
         return activityService.getById(id).getOwner();
     }
 
@@ -87,7 +87,7 @@ public class ActivityController {
     }
 
     @GetMapping("/activityId/{id}")
-    public Activity getById(@PathVariable("id") long id) throws ActivityNotFoundException {
+    public Activity getById(@PathVariable("id") int id) throws ActivityNotFoundException {
         return activityService.getById(id);
     }
 }
