+++ activity-microservice/src/main/java/nl/tudelft/sem/template/example/domain/ActivityNotFoundException.java
@@ -3,7 +3,7 @@ package nl.tudelft.sem.template.example.domain;
 public class ActivityNotFoundException extends Exception {
     private static final long serialVersionUID = 1;
 
-    public ActivityNotFoundException(Long id) {
+    public ActivityNotFoundException(int id) {
         super("Activity with id " + id + " not found");
     }
