+++ activity-microservice/src/main/java/nl/tudelft/sem/template/example/domain/ActivityService.java
@@ -57,7 +57,7 @@ public class ActivityService {
      * @param id
      * @param request
      */
-    public ResponseEntity editActivity(NetId netId, Long id, ActivityRequestModel request) throws UnauthorizedException, ActivityNotFoundException {
+    public ResponseEntity editActivity(NetId netId, int id, ActivityRequestModel request) throws UnauthorizedException, ActivityNotFoundException {
         Optional<Activity> activity = activityRepository.findById(id);
         if (activity.isPresent()) {
             Activity change = activity.get();
@@ -115,7 +115,7 @@ public class ActivityService {
         }
     }
 
-    public void deleteById(NetId netId, Long id) throws UnauthorizedException, ActivityNotFoundException {
+    public void deleteById(NetId netId, int id) throws UnauthorizedException, ActivityNotFoundException {
         Optional<Activity> activity = activityRepository.findById(id);
         if (activity.isPresent()) {
             if (activity.get().getOwner().getNetIdValue().equals(netId.getNetIdValue())) {
@@ -174,7 +174,7 @@ public class ActivityService {
         return result;
     }
 
-    public Activity getById(long id) throws ActivityNotFoundException {
+    public Activity getById(int id) throws ActivityNotFoundException {
         if (activityRepository.findById(id).isPresent()) {
             return activityRepository.findById(id).get();
         } else {
